Various improvements to TestCase class needed for unit testing

+++ scopehal/TestWaveformSource.cpp
@@ -82,7 +82,7 @@ TestWaveformSource::~TestWaveformSource()
 WaveformBase* TestWaveformSource::GenerateNoisySinewave(
 	float amplitude,
 	float startphase,
-	int64_t period,
+	float period,
 	int64_t sampleperiod,
 	size_t depth,
 	float noise_amplitude)
